♻️ Optimize ignores, make the "command" argument positional
Chokidar was throwing out thousands of events that were causing
performance issues. Rather than having the events be sent and then
filtering them out upstream, this adds some logic both to chokidar
as well as our event handler to make sure that events are culled
as close to the source as possible instead of being spammed out and
then filtered.

Also, opted for better argument handling. Rather than having to wrap
the command in a `--command` parameter, this just slurps everything
after the first thing that is not an option of watch itself

Instead of

```
watch --command "echo hello world"
```

```
watch echo hello world
```